Separate schedule list into separate component
