+++ frontend/awx/main/routes/useAwxInventoryRoutes.tsx
@@ -13,7 +13,6 @@ import { CreateSchedule } from '../../views/schedules/ScheduleForm';
 import { ScheduleDetails } from '../../views/schedules/SchedulePage/ScheduleDetails';
 import { SchedulePage } from '../../views/schedules/SchedulePage/SchedulePage';
 import { ScheduleRules } from '../../views/schedules/SchedulePage/ScheduleRules';
-import { Schedules } from '../../views/schedules/Schedules';
 import { AwxRoute } from '../AwxRoutes';
 import { InventoryJobTemplates } from '../../resources/inventories/InventoryPage/InventoryJobTemplates';
 import { InventoryHosts } from '../../resources/inventories/InventoryPage/InventoryHosts';
@@ -36,6 +35,7 @@ import { GroupDetails } from '../../resources/groups/GroupDetails';
 import { InventoryHostJobs } from '../../resources/inventories/inventoryHostsPage/InventoryHostJobs';
 import { InventoryHostFacts } from '../../resources/inventories/inventoryHostsPage/InventoryHostFacts';
 import { ResourceNotifications } from '../../resources/notifications/ResourceNotifications';
+import { SchedulesList } from '../../views/schedules/SchedulesList';
 
 export function useAwxInventoryRoutes() {
   const { t } = useTranslation();
@@ -108,7 +108,7 @@ export function useAwxInventoryRoutes() {
             {
               id: AwxRoute.InventorySourceSchedules,
               path: 'schedules',
-              element: <Schedules sublistEndpoint={awxAPI`/inventory_sources`} noHeader />,
+              element: <SchedulesList sublistEndpoint={awxAPI`/inventory_sources`} />,
             },
           ],
         },
